v0.1: user can insert, edit, delete movements and rid category based

+++ src/app/rid/rid.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild, OnInit, AfterViewInit } from '@angular/core';
-import { MatTableDataSource, MatTabChangeEvent, MatSort } from '@angular/material';
+import { MatTableDataSource, MatTabChangeEvent, MatSort, MatPaginator } from '@angular/material';
 
 import { Rid } from '../shared/rid';
 import { Info } from '../shared/info';
@@ -12,13 +12,15 @@ import { InfoService } from '../shared/info.service';
   styleUrls: ['./rid.component.css']
 })
 export class RidComponent implements OnInit, AfterViewInit {
-  displayedColumns = ['category', 'description', 'date', 'amount', 'action'];
+  displayedColumns = ['description', 'date', 'amount', 'verified', 'action'];
   dataSource = new MatTableDataSource();
   info: Info = {} as Info;
-  selectedTab: string;
+  selectedTab: string = 'Tutti';
   editing: Object = {};
   newRid: Rid = null;
+
   @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator) paginator: MatPaginator;
 
   constructor(private ridService: RidService, private infoService: InfoService) { }
   /**
@@ -27,10 +29,11 @@ export class RidComponent implements OnInit, AfterViewInit {
    */
   ngAfterViewInit() {
     this.dataSource.sort = this.sort;
+    this.dataSource.paginator = this.paginator;
     this.infoService.getInfo().then(info => {
       this.info = info;
     }).then(() => {
-      this.ridService.getAll({ category: this.info.categories[0].category }).then(rids => {
+      this.ridService.getAll().then(rids => {
         rids.forEach(rid => {
           this.editing[rid._id] = false;
         });
@@ -45,13 +48,23 @@ export class RidComponent implements OnInit, AfterViewInit {
     this.dataSource.filter = filterValue;
   }
   changedCategory(event: MatTabChangeEvent): void {
-    this.ridService.getAll({ category: this.info.categories[event.index].category }).then(rids => {
+    this.selectedTab = event.index === 0 ? 'Tutti' : this.info.categories[event.index - 1].category;
+    const filterCategory = this.selectedTab === 'Tutti' ? {} : { category: this.selectedTab };
+    this.ridService.getAll(filterCategory).then(rids => { // -1 because position 0 is all
       rids.forEach(rid => {
         this.editing[rid._id] = false;
       });
       this.dataSource.data = rids;
     });
     this.newRid = null;
+  }
+  verify(rid: Rid): void {
+    rid.verified = !rid.verified;
+    this.ridService.update(rid._id, rid).then(response => {
+
+    }).catch(error => {
+      alert(JSON.stringify(error, null, 2));
+    });
   }
   startEditing(rid: Rid): void {
     this.editing[rid._id] = true;
@@ -67,26 +80,27 @@ export class RidComponent implements OnInit, AfterViewInit {
     this.editing[rid._id] = false;
   }
   delete(rid: Rid): void {
-    // ask confirm before
-    this.ridService.delete(rid._id).then(() => {
-      this.dataSource.data = this.dataSource.data.filter(function (elem: Rid) {
-        return elem._id !== rid._id;
+    if (confirm('Eliminare elemeto?')) {
+      this.ridService.delete(rid._id).then(() => {
+        this.dataSource.data = this.dataSource.data.filter(function (elem: Rid) {
+          return elem._id !== rid._id;
+        });
+      }).catch(error => {
+        alert(JSON.stringify(error, null, 2));
       });
-    }).catch(error => {
-      alert(JSON.stringify(error, null, 2));
-    });
+    }
   }
   startNew(): void {
     this.newRid = new Rid();
     this.newRid.category = this.selectedTab;
     this.dataSource.data.unshift(this.newRid);
-    this.dataSource.filter = this.dataSource.filter;
+    this.dataSource.filter = '';
   }
   confirmNew(rid: Rid): void {
     const index: number = this.dataSource.data.findIndex(element => element === rid);
     this.ridService.create(rid).then(response => {
       this.dataSource.data[index] = response;
-      this.dataSource.filter = this.dataSource.filter;
+      this.dataSource.filter = '';
       this.newRid = null;
     }).catch(error => {
       alert(JSON.stringify(error, null, 2));
@@ -97,7 +111,7 @@ export class RidComponent implements OnInit, AfterViewInit {
     this.dataSource.data = this.dataSource.data.filter(function (elem: Rid) {
       return elem._id !== rid._id;
     });
-    this.dataSource.filter = this.dataSource.filter;
+    this.dataSource.filter = '';
   }
   validateAmount(event: any) {
     const pattern = /^[a-zA-Z]+$/;
